+++ lib/unused_files_analyzer.dart
@@ -1,7 +1,7 @@
-export 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/unused_files_analyzer.dart';
-export 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/unused_files_config.dart';
-
 export 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/models/unused_files_file_report.dart';
 
 export 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/reporters/reporters_list/console/unused_files_console_reporter.dart';
 export 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/reporters/reporters_list/json/unused_files_json_reporter.dart';
+
+export 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/unused_files_analyzer.dart';
+export 'package:dart_code_metrics/src/analyzers/unused_files_analyzer/unused_files_config.dart';